Renamed BracketsView storeies to BrackestViewDeploy
